refactor(hosts): move target/ into mtui/hosts/
Step 8 of the reorg plan, part 2 of 2: 'git mv mtui/target
mtui/hosts/target' as a whole directory. The eight files inside
target/ are single-class or near-single-class modules and have
been moved without further splits per plan section 3.6 (the
'split single-class files' anti-pattern guard).

Three classes of edits land in this commit:

1. Depth-shift of relative imports inside the moved tree.
   target/*.py modules used 'from ..foo' (2 dots) to reach
   mtui/foo. After moving to mtui/hosts/target/*.py, 2 dots now
   reach mtui/hosts/ — which is correct for sibling packages
   ('..connection', '..refhost') but wrong for the still-at-root
   packages (support, types, update_workflow, cli, etc.). Those
   imports get one extra dot: '..support.config' ->
   '...support.config'. target/parsers/system.py is one level
   deeper still, so '...types' -> '....types'.

   Net effect: 23 relative-import lines bumped across 8 files.
   The plan section 'Execution-time deviations from the plan'
   already documents this pattern.

2. Rewrite of mtui/ importers that name the old top-level path.
   16 sites across 11 files updated:
     mtui/cli/display.py:7
     mtui/update_workflow/hooks.py:18,19
     mtui/commands/{update,run,setrepo,_command,simpleset}.py
     mtui/test_reports/{testreport,pi_report,sl_report,
                       null_report,obs_report}.py
   All use the pattern '..target.X' -> '..hosts.target.X'.

3. Rewrite of tests/ — both 'from mtui.target.X' imports and
   the mock.patch / monkeypatch string targets that name modules
   inside target/. 87 substitutions of the literal 'mtui.target'
   -> 'mtui.hosts.target' across 26 test files (largest:
   test_hostgroup.py with 25 sites). The largest cluster of
   mock-string targets:
     'mtui.target.hostgroup.RunCommand' (and 15 siblings)
       -> 'mtui.hosts.target.hostgroup.RunCommand'  (19x)
     'mtui.target.target.parse_system' -> 'mtui.hosts.target.target.parse_system' (5x)
     'mtui.target.parsers.system.product' -> 'mtui.hosts.target.parsers.system.product' (2x)
     'mtui.target.actions.prompt_user' -> 'mtui.hosts.target.actions.prompt_user' (1x)

Logger names ('mtui.target', 'mtui.target.hostgroup',
'mtui.target.locks', 'mtui.target.operation',
'mtui.target.package_querier', 'mtui.target.repo_manager',
'mtui.target.reporter', 'mtui.targer.parsers.system' [sic]) are
kept at their original strings per the policy from commit 1be141b.
15 over-zealous caplog.at_level(logger='mtui.hosts.target...')
rewrites — produced by the bulk-sed of source imports — were
reverted to 'logger="mtui.target..."' so the WARNING / INFO /
DEBUG capture in tests/test_hostgroup.py, tests/test_target.py,
and tests/test_actions.py still resolves to the right loggers.
The two tests that briefly broke until the revert
(test_update_lock_logs_other_user_comment,
test_run_assertion_error_swallowed_with_debug_log) now pass.

Step 8 is complete. The plan's old top-level eradication test
('git grep -E ... connection|refhost|target ...') passes for
all three names.

Gates: ruff format --check, ruff check, ty check all green;
pytest = 989 passed; uv run python -m mtui --help works.
