+++ mtui/test_reports/sl_report.py
@@ -4,8 +4,8 @@ from logging import getLogger
 from typing import final
 
 from ..data_sources.gitea import Gitea
-from ..target import Target
-from ..target.hostgroup import HostsGroup
+from ..hosts.target import Target
+from ..hosts.target.hostgroup import HostsGroup
 from ..types import Product, RequestReviewID
 from .metadata_parsers import (
     JSONParser,
