+++ tests/conftest.py
@@ -79,7 +79,7 @@ def mock_connection():
 @pytest.fixture
 def mock_target(mock_config, mock_connection, mock_system):
     """A Target-like mock with realistic attributes and methods."""
-    from mtui.target import Target
+    from mtui.hosts.target import Target
 
     target = Target(mock_config, "host1.example.com")  # type: ignore[arg-type]
     target.connection = mock_connection
@@ -102,7 +102,7 @@ def mock_target(mock_config, mock_connection, mock_system):
 @pytest.fixture
 def mock_target_pair(mock_config, mock_system):
     """Two mock targets suitable for HostsGroup testing."""
-    from mtui.target import Target
+    from mtui.hosts.target import Target
 
     def make_target(hostname):
         t = Target(mock_config, hostname)  # type: ignore[arg-type]
