+++ tests/test_cmd_downgrade.py
@@ -8,8 +8,8 @@ from unittest.mock import MagicMock
 import pytest
 
 from mtui.commands.downgrade import Downgrade
+from mtui.hosts.target.hostgroup import HostsGroup
 from mtui.support.messages import NoRefhostsDefinedError, TestReportNotLoadedError
-from mtui.target.hostgroup import HostsGroup
 
 
 def _target(hostname="h1", state="enabled"):
