+++ tests/test_cmd_zypper.py
@@ -7,7 +7,7 @@ from argparse import Namespace
 from unittest.mock import MagicMock
 
 from mtui.commands.zypper import Install
-from mtui.target.hostgroup import HostsGroup
+from mtui.hosts.target.hostgroup import HostsGroup
 
 
 def _target(hostname):
